Reorganized the Pictures directory in the Atoms folder. Added the client head shot to the about page
