+++ src/components/Atoms/Pictures/AboutPhotos/aboutPhotos.css
@@ -1,0 +1,4 @@
+.aboutImg1 {
+    height: 500px;
+    width: auto;
+}
