+++ src/components/Atoms/Pictures/HomePhotos/heroImage.js
@@ -1,5 +1,5 @@
-import Image from "../../../Assets/Pictures/HeroImage/heroHome2.avif";
-import "./Pictures.css";
+import Image from "../../../../Assets/Pictures/HeroImage/heroHome2.avif";
+import "./heroImage.css";
 
 export default function HeroImage() {
     return (
