+++ src/components/Atoms/Pictures/Logo/logo.css
@@ -15,23 +15,9 @@
     width: auto;
 }
 
-.heroImageContainer {
-    height: 100%;
-    width: 100%;
-}
-
-.heroImage {
-    z-index: -1;
-    object-fit: cover;
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    top: 0;
-}
-
 @media screen and (max-width: 850px) {
     .logoImage {
-            height: auto;
-            width: 100vw;
-        }
+        height: auto;
+        width: 100vw;
+    }
 }
