+++ src/components/Molecules/AboutSection/AboutSection.js
@@ -1,0 +1,10 @@
+import "./aboutSection.css";
+import Headshot from "../../Atoms/Pictures/AboutPhotos/Headshot";
+
+export default function AboutSection() {
+    return (
+        <div className="aboutSection">
+            <Headshot />
+        </div>
+    );
+}
