+++ src/components/Molecules/AboutSection/aboutSection.css
@@ -1,0 +1,6 @@
+.aboutSection {
+    height: 100vh;
+    width: 90vw;
+    border: 2px solid black;
+    margin: 10% 0 10% 0;
+}
