+++ src/components/Molecules/Footer/Footer.js
@@ -1,6 +1,6 @@
 import "./footer.css";
 import Socials from "../SocialLinks/Socials";
-import LogoFooter from "../../Atoms/Pictures/LogoFooter";
+import LogoFooter from "../../Atoms/Pictures/Logo/LogoFooter";
 
 export default function Footer() {
   return (
