+++ src/components/Organisms/AboutPage/aboutContainer.css
@@ -1,0 +1,5 @@
+.aboutContainer {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
